pyproject.toml
[project]
name = "slackcmds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "slack-bolt>=1.23.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.3.5",
    "ruff>=0.11.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["slackcmds"]
python_files = ["test_*.py"]
addopts = "-v --cov=sologm --cov-report=term-missing"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

